CPM-541: refactor AddUuidSubscriberIntegration
