+++ tests/back/Pim/Enrichment/Integration/Product/MigrateToUuidCommandIntegration.php
@@ -8,6 +8,7 @@ use Akeneo\Pim\Automation\DataQualityInsights\Application\ProductEvaluation\Eval
 use Akeneo\Pim\Enrichment\Bundle\Command\MigrateToUuid\MigrateToUuidAddTriggers;
 use Akeneo\Pim\Enrichment\Bundle\Command\MigrateToUuid\MigrateToUuidStep;
 use Akeneo\Pim\Enrichment\Product\API\Command\UpsertProductCommand;
+use AkeneoTest\Pim\Enrichment\Integration\Product\UuidMigration\AbstractMigrateToUuidTestCase;
 use PHPUnit\Framework\Assert;
 
 final class MigrateToUuidCommandIntegration extends AbstractMigrateToUuidTestCase
